push

+++ app/Http/Controllers/TaskController.php
@@ -20,37 +20,31 @@ class TaskController extends Controller
 {
     public function index()
     {
-        $tasks = Task::where('user_id', Auth::id())->paginate(5);
-//        $minmax_array = Product::selectRaw('MIN(price) as min_price, MAX(price) as max_price')->get()->all();
-        $minmax_array = Product::selectRaw('
-            MIN(price) as min_price,
-            MAX(price) as max_price
-        ')->first();
-//        print_r($minmax_array);
-        echo $minmax_array->min_price .  '  ' .  $minmax_array->max_price;
-//        $maxprice = $minmax_array['max_price'];
-//        $minprice = $minmax_array['min_price'];
+        $string_with_256_symbols = Str::random(256);
 //        $faker = new Faker;
 //        $string_with_256_symbols = $faker->text(110);
-
-//        $string_with_256_symbols = Str::random(256);
-//
 //        Debugbar::info($string_with_256_symbols);    // Debugbar usage example
 //        Debugbar::error('Error!');
 //        Debugbar::warning('Watch out…');
 //        Debugbar::addMessage('Another message', 'mylabel');
-//
-//        return view('tasks.tasks', compact('tasks', 'minmax_array'));
+
+        $tasks = Task::with('ratings', 'user')->paginate(10);
+//        $tasks = Task::with('ratings')->where('user_id', Auth::id())->paginate(5);
+//        $tasks = Task::where('user_id', Auth::id())->paginate(5);
+        return view('tasks.tasks', compact('tasks', 'string_with_256_symbols'));
     }
 
     public function task_info($task_ID)
     {
+
         $task = Task::find($task_ID);
         return view('tasks.task_info', ['task' => $task]);
     }
 
     public function post(TaskPostRequest $request)
     {
+        $this->authorize('create_task');
+
         $user = Auth::user();
         $validated = $request->validated();
         $task = Task::create([
@@ -73,6 +67,8 @@ class TaskController extends Controller
 
     public function update_page($task_ID=null)
     {
+        $this->authorize('update_task');
+
         if (isset($task_ID)) {
             $found_task = Task::find($task_ID);
         }
@@ -84,25 +80,39 @@ class TaskController extends Controller
 
     public function update(TaskPostRequest $request, $found_task)
     {
+        $this->authorize('update_task');
+
         $user = Auth::user();
         $validated = $request->validated();
-        $task = Task::find($found_task)->update([
-            'name' => $validated['name'],
-            'details' => $validated['details'],
-            'status' => $status = isset($validated['status']) ? $validated['status'] : null,
-            'priority' => $validated['priority'],
-            'start_time' => $validated['start_time'],
-            'finish_time' => $validated['finish_time'],
-            'time_spent' => $validated['time_spent'],
-        ]);
         if (isset($validated['rating'])) {
-            $task->rate($validated['rating'], $user);
+            Task::find($found_task)->rate($validated['rating'], $user)->update([
+                'name' => $validated['name'],
+                'details' => $validated['details'],
+                'status' => $status = isset($validated['status']) ? $validated['status'] : null,
+                'priority' => $validated['priority'],
+                'start_time' => $validated['start_time'],
+                'finish_time' => $validated['finish_time'],
+                'time_spent' => $validated['time_spent'],
+            ]);
+        }
+        else {
+            Task::find($found_task)->update([
+                'name' => $validated['name'],
+                'details' => $validated['details'],
+                'status' => $status = isset($validated['status']) ? $validated['status'] : null,
+                'priority' => $validated['priority'],
+                'start_time' => $validated['start_time'],
+                'finish_time' => $validated['finish_time'],
+                'time_spent' => $validated['time_spent'],
+            ]);
         }
         return redirect()->route('tasks_main_page');
     }
 
     public function delete($task_ID)
     {
+        $this->authorize('delete_task');
+
         $found_task = Task::find($task_ID);
 
         $found_task->delete();
